requirements.txt
# Основные зависимости:
opencv-python>=4.5
deepface>=0.0.75
telebot>=0.0.5
Pillow>=9.0
numpy>=1.21
scikit-learn>=1.0
NumPy
datetime